backend/requirements.txt
python-dotenv==1.0.0
google-generativeai
requests==2.31.0
fastapi==0.110
pydantic==2.7
langgraph
tavily-python
uvicorn[standard]==0.29
openai==1.30
httpx==0.27
apify-client
groq
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-magic==0.4.27
aiofiles==23.2.1
websockets==12.0
Pillow==10.0.0
streamlit>=1.28.0
colorsys==0.1.10

# Enhanced AI-powered vectorization dependencies
rembg>=2.0.50
opencv-python>=4.8.0
scikit-image>=0.21.0
scikit-learn>=1.3.0
scipy>=1.11.0
numpy>=1.24.0

onnxruntime==1.15.1
onnxruntime-gpu==1.15.1